requirements.txt
torch>=1.9.0
transformers>=4.24.0
diffusers>=0.9.0
streamlit>=1.3.0
Pillow>=8.4.0